+++ src/main/java/org/folio/uk/integration/roles/model/UserRoles.java
@@ -1,0 +1,13 @@
+package org.folio.uk.integration.roles.model;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+import lombok.Data;
+
+@Data
+public class UserRoles {
+
+  private UUID userId;
+  private List<UUID> roleIds = new ArrayList<>();
+}